indicators crate reorganization
